+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.2] - 2024-11-30
+### Added
+- Bearbeitungsmöglichkeit der Bestätigungs-Email auf der Einstellungsseite im `wp-admin-bereich`.
+- Live-Vorschau der Bestätigungs-Email
+- Versand einer Test-Bestätigungs-Email
+
+---
+
 ## [1.1] - 2024-11-23
 ### Added
 - Einstellungsseite im WordPress-Admin-Bereich für die Konfiguration des Plugins.